+++ mgl-webapp/src/main/java/com/organization/mvcproject/controller/GameController.java
@@ -8,6 +8,8 @@ import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -18,31 +20,12 @@ import com.organization.mvcproject.model.Review;
 import com.organization.mvcproject.service.GameService;
 
 @Controller
-public class GameAndReviewController {
+public class GameController {
 
 	@Autowired
 	private GameService gameService;
-
-	@RequestMapping(value = "/", method = RequestMethod.GET)
-	public String home() {
-		return "index";
-	}
 	
-	@RequestMapping(value = "/review", method = RequestMethod.GET)
-	public ModelAndView review() {
-		return new ModelAndView("reviewCreatePage", "command", new Review());
-	}
-
-	@RequestMapping(value = "/addReview", method = RequestMethod.POST)
-	public ModelAndView addReview(Review review, ModelMap model) {
-		if(review.getAuthor().equals("")) {
-			review.setAuthor("anonymous");
-		}
-		return new ModelAndView("reviewDetailPage", "submittedReview", review);
-	}
-
-	
-	@RequestMapping(value = "/games", method = RequestMethod.GET)
+	@GetMapping("games")
 	public ModelAndView game() {
 		return new ModelAndView("gamesPage", "command", new Game());
 	}
@@ -51,12 +34,12 @@ public class GameAndReviewController {
 	 * TODO 2.0 (Separation of concerns) consider moving all controller endpoints that return a ResponseEntity into a @RestController.
 	 */
 	
-	@RequestMapping(value = "/game", method = RequestMethod.GET)
+	@GetMapping("game")
 	public ResponseEntity<List<Game>> fetchAllGames() {
 		return new ResponseEntity<List<Game>>(gameService.retrieveAllGames(), HttpStatus.OK);
 	}
 
-	@RequestMapping(value = "/game", method = RequestMethod.POST, consumes = MediaType.APPLICATION_JSON_VALUE)
+	@PostMapping(value = "/game", consumes = MediaType.APPLICATION_JSON_VALUE)
 	public ResponseEntity<Void> createGame(@RequestBody Game game) {
 		gameService.saveGame(game);
 		return new ResponseEntity<Void>(HttpStatus.CREATED);
